pyproject.toml
[project]
name = "langgraph-lectures"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.2.0",
    "bs4>=0.0.2",
    "chromadb>=1.0.16",
    "faiss-cpu>=1.11.0.post1",
    "fastapi>=0.116.1",
    "fitz>=0.0.1.dev2",
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-groq>=0.3.7",
    "langchain-huggingface>=0.3.1",
    "langchain-ollama>=0.3.6",
    "langchain-text-splitters>=0.3.9",
    "langserve>=0.3.1",
    "pymupdf>=1.26.3",
    "pypdf>=5.9.0",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.1",
    "sentence-transformers>=5.1.0",
    "sse-starlette>=3.0.2",
    "starlette>=0.47.2",
    "tqdm>=4.67.1",
    "uvicorn>=0.35.0",
    "wikipedia>=1.4.0",
]
